pyproject.toml
[project]
name = "do-again-list"
version = "0.0.1"
requires-python = ">=3.12"
description="An un-todo list for tracking the last time you did things that you want to do more/less frequently/regularly"
authors = [
    {name = "Chad Spratt"},
]
readme = "README.md"
# license = "AGPL-3.0-or-later"
# license-files = ["LICENSE.txt"]
keywords = ["productivity", "habit management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    # "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Utilities",
]
dependencies = [
    "django>=5.0.0",
    "django-appconf>=1.2.0",
    "djangorestframework>=3.16.1",
    "django-filter>=25.2",
]


[project.urls]
repository = "https://github.com/chadspratt/do_again_list"

[dependency-groups]
dev = [
    "drf-spectacular>=0.29.0",
    "django-cors-headers>=4.9.0",
    "ruff>=0.15.1",
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.8.0",
    "tox>=4.11.3",
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "ty>=0.0.18",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "do_again_list"
module-root = ""
targets = ["do_again_list"]

[tool.ruff]
exclude = ["migrations"]

[tool.ruff.lint]
ignore = ["E402", "D105", "D106"]
select = ["E4", "E7", "E9", "F", "I", "D1"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["test_project"]

[tool.ty.src]
exclude = [
    "test_project/**",
]
